Add simple schema validation
This commit adds a very simple validation test for the schemas.
Validation is performed against json-schema draft 4 meta-schema.
Note that this validates only the single file and does not
follow refs. So while refs are syntactically correct they
may still be incorrect and pointing to the wrong things.

Requires jsonschema to be installed in the current python env.