Adding sample events to validate against
These samples were generated by stacktach-notigen. Only one
sample of each event type has been included, but the naming
scheme allows for multiples to allow adding multiples of the
same type for testing variants.

Some loosening of the schemas was required to get these to
validate. I believe some of that was caused by the way the
events were generated and dumped by json, e.g. causing numeric
strings to be emitted as actual numbers. Unsure how this will
compare to what actually is generated in a live system.